+++ src/background/index.ts
@@ -1,3 +1,5 @@
+import { initRuntimeLog } from "../utils/runtimeLog";
 import { setupMessageListeners } from "./messageListener";
 
+initRuntimeLog().catch(() => {});
 setupMessageListeners();