+++ test/jest/__mocks__/webextension-polyfill.ts
@@ -22,6 +22,10 @@ const browser = {
       remove: jest.fn().mockResolvedValue(undefined),
       clear: jest.fn().mockResolvedValue(undefined),
     },
+    onChanged: {
+      addListener: jest.fn(),
+      removeListener: jest.fn(),
+    },
   },
   
   tabs: {